+++ NEWS.md
@@ -28,7 +28,7 @@
 
 * Fix issues of `TimeArray` column names copying. (#418)
 
-* `ta[]` throws `BoundsError` now. (#420)
+* `timearray[]` throws `BoundsError` now. (#420)
 
   ```julia
   julia> cl[]
@@ -37,6 +37,15 @@
   ```
 
 * `Tables.jl` interface integration. (#382)
+
+* 2D `getindex` supports. (#423)
+
+  ```julia
+  ohlc[1:42,   [:High, :Low]]
+  ohlc[42:end, [:High, :Low]]
+  ohlc[:,      [:High, :Low]]
+  ohlc[42,     [:High, :Low]]
+  ```
 
 
 ### 0.15.0